fixed big double amount bug

+++ app/src/main/java/com/androidcollider/easyfin/utils/FormatUtils.java
@@ -2,20 +2,18 @@ package com.androidcollider.easyfin.utils;
 
 
 import android.support.annotation.NonNull;
-
 import java.text.DecimalFormat;
-import java.text.NumberFormat;
-import java.util.Locale;
+
 
 public class FormatUtils {
 
     @NonNull
     public static String doubleFormatter (double number, String format, int precise){
-        //NumberFormat numberFormat = NumberFormat.getNumberInstance(Locale.GERMAN);
+
         DecimalFormat dfRate = new DecimalFormat(format);
         precise = 10^precise;
         number = number*precise;
-        int i = (int) Math.round(number);
+        long i = Math.round(number);
         double result = (double) i/precise;
         String s = dfRate.format(result);
